updated Collector app functionalities and corrections, Major UI theming applied.

+++ mb40mktg-android/app/src/main/java/com/android/itproj/mb40marketing/model/TransactionModel.java
@@ -47,7 +47,7 @@ public class TransactionModel {
     @SerializedName("payment")
     @Getter
     @Setter
-    private int payment;
+    private double payment;
 
     @SerializedName("created_at")
     @Getter @Setter
